+++ database/migrations/2022_09_12_000015_create_content_category_content_page_pivot_table.php
@@ -8,11 +8,13 @@ class CreateContentCategoryContentPagePivotTable extends Migration
 {
     public function up()
     {
-        Schema::create('content_category_content_page', function (Blueprint $table) {
-            $table->unsignedBigInteger('content_page_id');
-            $table->foreign('content_page_id', 'content_page_id_fk_7304923')->references('id')->on('content_pages')->onDelete('cascade');
-            $table->unsignedBigInteger('content_category_id');
-            $table->foreign('content_category_id', 'content_category_id_fk_7304923')->references('id')->on('content_categories')->onDelete('cascade');
-        });
+        if (Schema::hasTable('content_category_content_page') === false) {
+            Schema::create('content_category_content_page', function (Blueprint $table) {
+                $table->unsignedBigInteger('content_page_id');
+                $table->foreign('content_page_id', 'content_page_id_fk_7304923')->references('id')->on('content_pages')->onDelete('cascade');
+                $table->unsignedBigInteger('content_category_id');
+                $table->foreign('content_category_id', 'content_category_id_fk_7304923')->references('id')->on('content_categories')->onDelete('cascade');
+            });
+        }
     }
 }
